Core remove legacy patch (#1047)
